Refactored the grpc server to organize functionality by handler.
